Refactor and controller creation
