+++ src/middleware/errorHandler.js
@@ -7,4 +7,4 @@ const errorHandler = (err, req, res, next) => {
   res.status(500).send(err.message);
 };
 
-module.exports = { errorHandler };
+module.exports = errorHandler;
