+++ src/middleware/logEvents.js
@@ -12,12 +12,12 @@ const logEvents = async (message, fileName) => {
   console.log(logItem);
 
   try {
-    if (!fs.existsSync(path.join(__dirname, "..", "logs"))) {
-      fsPromises.mkdir(path.join(__dirname, "..", "logs"));
+    if (!fs.existsSync(path.join(__dirname, "..", "..", "logs"))) {
+      fsPromises.mkdir(path.join(__dirname, "..", "..", "logs"));
     }
 
     await fsPromises.appendFile(
-      path.join(__dirname, "..", "logs", fileName),
+      path.join(__dirname, "..", "..", "logs", fileName),
       logItem
     );
   } catch (error) {
