+++ src/routes/root.routes.js
@@ -1,18 +1,10 @@
 const { Router } = require("express");
 const path = require("node:path");
 
-const rootRoute = Router();
+const rootRoutes = Router();
 
-rootRoute.get("^/$|/index(.html)?", (req, res) => {
+rootRoutes.get("^/$|/index(.html)?", (req, res) => {
   res.sendFile(path.join(__dirname, "..", "..", "views", "index.html"));
 });
 
-rootRoute.get("/new-page(.html)?", (req, res) => {
-  res.sendFile(path.join(__dirname, "..", "..", "views", "new-page.html"));
-});
-
-rootRoute.get("/old-page(.html)?", (req, res) => {
-  res.redirect(301, "/new-page.html");
-});
-
-module.exports = rootRoute;
+module.exports = rootRoutes;
