--- views/new-page.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link href="css/style.css" rel="stylesheet" />
-    <title>New page</title>
-  </head>
-
-  <body>
-    <h1>New page...</h1>
-    <img src="../img/avatar.jpg" alt="avatar" />
-  </body>
-</html>
