--- views/subdir/test.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link href="css/style.css" rel="stylesheet" />
-    <title>Subdir | Testing</title>
-  </head>
-
-  <body>
-    <h1>Testing Routes page...</h1>
-  </body>
-</html>
